+++ package.json
@@ -32,7 +32,8 @@
     "jsonwebtoken": "^8.5.1",
     "moment": "^2.24.0",
     "mongoose": "^5.8.5",
-    "morgan": "~1.9.1"
+    "morgan": "~1.9.1",
+    "react-transition-group": "^4.3.0"
   },
   "devDependencies": {
     "eslint": "^6.8.0",
